Redesign according to the KISS principle
Such a simple site as this does not need 100K of JavaScript. It does not need
12K of CSS. It does not require a build process. Keep it simple!

Instead of the inaccessible and slow “parallax” effect, you now always scroll
down to see the contact information and projects. There’s a simple “reveal”
effect instead. That keeps it consistent on both desktop and mobile.